Enemy behaviours

+++ gameDev-GroupH/Assets/Scripts/SceneShift/MenuController.cs
@@ -21,7 +21,7 @@ public class MenuController : MonoBehaviour
         saveState.Death.Clear();
         saveState.ClueCount = 0;
         saveState.SaveLocation((float)-115.4, 1, (float)-65.9);
-        saveState.SavePlayerHealth(new float [] { ply.getTotalMaxHP(), ply.getTotalMaxHP(), ply.getTotalMaxHP(), ply.getTotalMaxHP() });
+        saveState.SavePlayerHealth(new int [] { ply.getTotalMaxHP(), ply.getTotalMaxHP(), ply.getTotalMaxHP(), ply.getTotalMaxHP() });
     }
 
     //starts the game when play button is pressed
